1_gather_and_generate_training_data/requirements.txt
datasets==3.0.1
scipy==1.14.1
numpy==2.0.2
soundfile==0.12.1
librosa==0.10.2
progressbar==2.5
gitpython==3.1.43
torch==2.4.1
torchaudio==2.4.1
piper-tts==1.2.0
webrtcvad==2.0.10
azure-cognitiveservices-speech==1.40.0